requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohttp 
aiosignal 
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apex 
argcomplete==1.12.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asciitree 
asttokens==2.4.1
astunparse==1.6.3
async-timeout 
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==7.1.2
cloudpathlib==0.16.0
cloudpickle 
cmake==3.29.2
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
crcmod==1.7
crypto==1.4.1
cuda-python 
cuml 
cupy-cuda12x 
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dask 
dask-cuda 
dask-expr 
datasets==2.20.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.2
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distributed 
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
email_validator==2.2.0
entrypoints 
et-xmlfile==1.1.0
exceptiongroup==1.2.1
execnet==2.1.1
executing==2.0.1
expecttest==0.1.3
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners 
fastjsonschema==2.19.1
fastrlock 
filelock==3.14.0
flash-attn 
fonttools==4.51.0
frozenlist 
fsspec==2024.3.1
galaxy-fds-sdk==1.4.41
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.2
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gpustat==1.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.30.2
hydra-core==1.3.2
hypothesis==5.35.1
idna==3.7
igraph==0.11.4
ijson==3.3.0
importlib_metadata 
iniconfig==2.0.0
intel-openmp==2021.4.0
interegular==0.3.3
ipykernel==6.29.4
ipython==8.21.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib 
json5==0.9.25
jsonlines==4.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-tensorboard 
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab==2.3.2
jupyterlab-server==1.2.0
jupyterlab_pygments==0.3.0
jupytext==1.16.1
kiwisolver==1.4.5
kvikio 
langcodes==3.4.0
language_data==1.2.0
lark==1.1.9
lazy_loader==0.4
librosa==0.10.1
lightning==2.5.1.post0
lightning-thunder==0.2.0.dev0
lightning-utilities==0.11.2
llvmlite==0.42.0
lm-format-enforcer==0.10.1
lmdb==1.5.1
locket 
looseversion==1.3.0
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==5.1.0
more-itertools==10.3.0
motor==3.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict 
multiprocess==0.70.16
murmurhash==1.0.10
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
notebook==6.4.10
numba==0.59.1
numcodecs 
numpy==1.24.4
nvfuser==0.2.0a0+0ff5802
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-frontend 
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.37.1
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvimgcodec-cu12==0.2.0.7
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
nvitop==1.4.2
nvtx 
nx-cugraph 
oauthlib==3.2.2
omegaconf==2.3.0
onnx 
openai==1.35.4
opencv 
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
outlines==0.0.46
packaging==25.0
pandas 
pandocfilters==1.5.1
parso==0.8.4
partd 
peft==0.15.2
pexpect==4.9.0
Pillow 
platformdirs==4.2.1
pluggy==1.5.0
ply 
polygraphy==0.49.10
pooch==1.8.1
portalocker==2.10.0
preshed==3.0.9
prettytable==3.10.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pybind11_global==2.12.0
pycocotools 
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyGObject==3.42.1
pylibcugraph 
pylibcugraphops 
pylibraft 
pylibwholegraph 
pymongo==4.8.0
pynvjitlink 
pynvml 
pyparsing==3.1.2
pyrsistent==0.18.1
pytest==8.1.1
pytest-flakefinder==1.1.0
pytest-rerunfailures==14.0
pytest-shard==0.1.2
pytest-xdist==3.6.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-hostlist==1.23.0
python-multipart==0.0.9
pytorch-lightning==2.5.1.post0
pytorch-quantization==2.1.2
pytorch-triton 
pytz 
PyYAML==6.0.1
pyzmq==26.0.3
rados==2.0.0
raft-dask 
rapids-dask-dependency 
ray==2.30.0
rbd==2.0.0
referencing==0.35.1
regex==2024.4.28
requests==2.32.3
requests-oauthlib==2.0.0
rich 
rmm 
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy 
Send2Trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.12.0
setproctitle==1.3.3
shapely==2.1.0
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx_glpi_theme==0.6
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tblib 
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorrt 
terminado==0.18.1
texttable==1.7.0
thinc==8.2.3
threadpoolctl==3.5.0
thriftpy2 
tiktoken==0.7.0
timm==1.0.12
tinycss2==1.3.0
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.1
toolz 
torch==2.4.1
torch-tb-profiler==0.4.3
torch-tensorrt 
torchaudio==2.4.1
torchmetrics==1.7.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.9.0
transformers==4.45.2
treelite 
triton==3.0.0
typer==0.12.3
types-dataclasses==0.6.6
typing_extensions==4.11.0
tzdata 
ubuntu-pro-client==8001
ucx-py 
ujson==5.10.0
urllib3 
uv==0.5.29
uvicorn==0.30.1
uvloop==0.19.0
vertexai==1.71.1
wandb==0.17.6
wasabi==1.1.2
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.2.3
websockets==12.0
Werkzeug==3.0.2
xdoctest==1.0.2
xformers 
xgboost 
xxhash==3.4.1
yarl 
zarr 
zict 
zipp 
